refactored MS-graph response deserialization types out to their own files
