+++ src/SelfService/Infrastructure/BackgroundJobs/DeserializationTypes/AuthtokenReponse.cs
@@ -1,0 +1,17 @@
+using System.Text.Json.Serialization;
+namespace SelfService.Infrastructure.BackgroundJobs;
+
+public class AuthTokenResponse
+{
+    [JsonPropertyName("token_type")]
+    public string? TokenType { get; set; }
+
+    [JsonPropertyName("expires_in")]
+    public int ExpiresIn { get; set; }
+
+    [JsonPropertyName("ext_expires_in")]
+    public int ExternalExpiresIn { get; set; }
+
+    [JsonPropertyName("access_token")]
+    public string? AccessToken { get; set; }
+}
